Observe GL context loss in ShaderEffectItem
Recreate the shader program as required when the GL context
changes.  Also delete the shader program when the ShaderEffectItem
is deactivated, to save graphics memory.

Previously, this was done only for Symbian platforms, and in a
suboptimal fashion.  This patch implements the behaviour for all
platforms in a more optimal manner.

Signed-off-by: Chris Adams <chris.adams@jollamobile.com>