Ensure QDeclarativeTextLayout uses the correct pen
When a styled text specifies a font region with some color set,
QDeclarativeTextLayout used to revert to using a default-constructed
(black-colored) pen for text occurring after the </font> tag, rather
than the "default" (property-defined) color.

This commit ensures that the correct pen is used, by setting the pen
of the "inert" painter to be an invalid color rather than a default
constructed pen.